closing the dialog via javascript
