+++ app/components/data_imports/linelist_import_dialog/v2/component.html.erb
@@ -22,9 +22,7 @@
     <%= helpers.turbo_frame_tag "import_metadata_dialog_alert" %>
 
     <%= tag.div id: "import-metadata-dialog-content", data: linelist_data do %>
-      <%= form_for(:file_import, url: "#", method: :get, data: {
-        action: "submit-end->viral--dialog#close",
-      }) do |form| %>
+      <%= form_for(:file_import, url: "#", method: :get) do |form| %>
         <%= render partial: 'shared/samples/metadata/file_imports/form_fields', locals: {
           broadcast_target: @broadcast_target,
           controller: "linelist-import",
